+++ src/frontend/Komponenter/Behandling/BehandlingContainer.tsx
@@ -9,7 +9,6 @@ import { BehandlingProvider, useBehandling } from '../../App/context/BehandlingC
 import DataViewer from '../../Felles/DataViewer/DataViewer';
 import PersonHeaderComponent from '../../Felles/PersonHeader/PersonHeader';
 import { GodkjennEndringer } from './Endringer/GodkjennEndringer';
-import { BrevmottakereModalForBehandling } from './Brevmottakere/BrevmottakereModalForBehandling';
 import { Behandling } from '../../App/typer/fagsak';
 import { IPersonopplysninger } from '../../App/typer/personopplysninger';
 import { HenleggModal } from './Henleggelse/HenleggModal';
@@ -79,10 +78,6 @@ const BehandlingContent: FC<{
                     <InfostripeUtestengelse utestengelser={utestengelser} />
                     <BehandlingRoutes />
                     <GodkjennEndringer behandling={behandling} />
-                    <BrevmottakereModalForBehandling
-                        behandlingId={behandling.id}
-                        personopplysninger={personopplysninger}
-                    />
                     <HenleggModal behandling={behandling} />
                 </InnholdWrapper>
                 <HøyreMenyWrapper åpenHøyremeny={åpenHøyremeny}>
