+++ src/frontend/Komponenter/Behandling/Brev/Brevmeny.tsx
@@ -32,6 +32,7 @@ const StyledBrevMeny = styled.div`
     display: flex;
     flex-direction: column;
     min-width: 450px;
+    margin-top: 2rem;
 `;
 
 const datasett = 'ef-brev';
